nanohttpd: refactor: move HTTPSession to a new file
